split utils in two files
